+++ config/crd/patches/cainjection_in_watchers.yaml
