+++ config/rbac/watcher_editor_role.yaml
