+++ config/rbac/watcher_viewer_role.yaml
