+++ config/samples/k1_v1beta1_watcher.yaml
